Fix momentum scrolling in iOS
When `overflow` is set to `auto`, iOS does not enable momentum
scrolling for the overflowed content unless it is explicitly set
with `-webkit-overflow-scrolling: touch`. This was most evident on
the examples page on an iPad. I added that rule to the body, which
maintains its 100% height and auto overflow, and momentum scrolling
works as expected.